backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pymongo==4.6.2
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
motor==3.3.2
python-multipart==0.0.9
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
websockets
requests
